+++ packages/twenty-server/src/engine/api/graphql/workspace-schema-builder/enums/connection-type-definition-kind.enum.ts
@@ -1,0 +1,4 @@
+export enum ConnectionTypeDefinitionKind {
+  Edge = 'Edge',
+  PageInfo = 'PageInfo',
+}
