+++ packages/twenty-server/src/engine/api/graphql/workspace-schema-builder/gql-type-generators/aggregation-type.generator.ts
