+++ packages/twenty-server/src/engine/api/graphql/workspace-schema-builder/gql-type-generators/composite-field-object-type.generator.ts
@@ -1,21 +1,20 @@
 import { Injectable, Logger } from '@nestjs/common';
 
-import { type GraphQLFieldConfigMap, GraphQLObjectType } from 'graphql';
+import { GraphQLObjectType } from 'graphql';
 import { FieldMetadataType } from 'twenty-shared/types';
 
-import { StoredGqlType } from 'src/engine/api/graphql/workspace-schema-builder/interfaces/stored-gql-type.interface';
+import { StoredObjectType } from 'src/engine/api/graphql/workspace-schema-builder/interfaces/stored-gql-type.interface';
 import { type WorkspaceBuildSchemaOptions } from 'src/engine/api/graphql/workspace-schema-builder/interfaces/workspace-build-schema-options.interface';
 import { type CompositeType } from 'src/engine/metadata-modules/field-metadata/interfaces/composite-type.interface';
 
 import { ObjectTypeDefinitionKind } from 'src/engine/api/graphql/workspace-schema-builder/enums/object-type-definition-kind.enum';
 import { FieldObjectTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/gql-type-generators/field-object-type.generator';
+import { GraphQLOutputTypeFieldConfigMap } from 'src/engine/api/graphql/workspace-schema-builder/types/graphql-field-config-map.types';
 import { computeCompositeFieldEnumTypeKey } from 'src/engine/api/graphql/workspace-schema-builder/utils/compute-stored-gql-type-key-utils/compute-composite-field-enum-type-key.util';
 import { computeCompositeFieldObjectTypeKey } from 'src/engine/api/graphql/workspace-schema-builder/utils/compute-stored-gql-type-key-utils/compute-composite-field-object-type-key.util';
 import { isEnumFieldMetadataType } from 'src/engine/metadata-modules/field-metadata/utils/is-enum-field-metadata-type.util';
 import { isMorphOrRelationFieldMetadataType } from 'src/engine/utils/is-morph-or-relation-field-metadata-type.util';
 import { pascalCase } from 'src/utils/pascal-case';
-
-export interface StoredObjectType extends StoredGqlType<GraphQLObjectType> {}
 
 @Injectable()
 export class CompositeFieldObjectTypeGenerator {
@@ -44,11 +43,8 @@ export class CompositeFieldObjectTypeGenerator {
   private generateFields(
     compositeType: CompositeType,
     options: WorkspaceBuildSchemaOptions,
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  ): GraphQLFieldConfigMap<any, any> {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const fields: GraphQLFieldConfigMap<any, any> = {};
+  ): GraphQLOutputTypeFieldConfigMap {
+    const fields: GraphQLOutputTypeFieldConfigMap = {};
 
     for (const property of compositeType.properties) {
       // Relation fields are not supported in composite types
