+++ packages/twenty-server/src/engine/api/graphql/workspace-schema-builder/gql-type-generators/gql-type.generator.ts
@@ -3,16 +3,16 @@ import { Injectable, Logger } from '@nestjs/common';
 import { WorkspaceBuildSchemaOptions } from 'src/engine/api/graphql/workspace-schema-builder/interfaces/workspace-build-schema-options.interface';
 import { type CompositeType } from 'src/engine/metadata-modules/field-metadata/interfaces/composite-type.interface';
 
-import { CompositeFieldInputTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/factories/composite-input-type-definition.factory';
-import { CompositeFieldObjectTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/factories/composite-object-type-definition.factory';
-import { ConnectionObjectTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/factories/connection-type-definition.factory';
-import { EdgeObjectTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/factories/edge-type-definition.factory';
-import { EnumFieldTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/factories/enum-type-definition.factory';
-import { ExtendedObjectMetadataObjectTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/factories/extend-object-type-definition-v2.factory';
-import { ObjectMetadataObjectTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/factories/object-type-definition.factory';
-import { RelationConnectInputTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/factories/relation-connect-input-type-definition.factory';
 import { CompositeFieldEnumTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/gql-type-generators/composite-field-enum-type.generator';
+import { CompositeFieldInputTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/gql-type-generators/composite-field-input-type.generator';
+import { CompositeFieldObjectTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/gql-type-generators/composite-field-object-type.generator';
+import { ConnectionObjectTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/gql-type-generators/connection-object-type.generator';
+import { EdgeObjectTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/gql-type-generators/edge-object-type.generator';
+import { EnumFieldTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/gql-type-generators/enum-field-type.generator';
+import { ExtendedObjectMetadataObjectTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/gql-type-generators/extended-object-metadata-object-type.generator';
 import { ObjectMetadataInputTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/gql-type-generators/object-metadata-input-type.generator';
+import { ObjectMetadataObjectTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/gql-type-generators/object-metadata-object-type.generator';
+import { RelationConnectInputTypeGenerator } from 'src/engine/api/graphql/workspace-schema-builder/gql-type-generators/relation-connect-input-type.generator';
 import { TypeDefinitionsStorage } from 'src/engine/api/graphql/workspace-schema-builder/storages/type-definitions.storage';
 import { objectContainsMorphRelationField } from 'src/engine/api/graphql/workspace-schema-builder/utils/object-contains-morph-relation-field.util';
 import { objectContainsRelationField } from 'src/engine/api/graphql/workspace-schema-builder/utils/object-contains-relation-field';
