+++ packages/twenty-server/src/engine/api/graphql/workspace-schema-builder/interfaces/stored-gql-type.interface.ts
@@ -1,4 +1,17 @@
+import {
+  type GraphQLEnumType,
+  type GraphQLInputObjectType,
+  type GraphQLObjectType,
+} from 'graphql';
+
 export interface StoredGqlType<T> {
   key: string;
   type: T;
 }
+
+export interface StoredInputType
+  extends StoredGqlType<GraphQLInputObjectType> {}
+
+export interface StoredObjectType extends StoredGqlType<GraphQLObjectType> {}
+
+export interface StoredEnumGqlType extends StoredGqlType<GraphQLEnumType> {}
