+++ packages/twenty-server/src/engine/api/graphql/workspace-schema-builder/storages/type-definitions.storage.ts
@@ -7,11 +7,14 @@ import {
 } from 'graphql';
 import { type FieldMetadataType } from 'twenty-shared/types';
 
+import {
+  StoredEnumGqlType,
+  StoredInputType,
+  StoredObjectType,
+} from 'src/engine/api/graphql/workspace-schema-builder/interfaces/stored-gql-type.interface';
+
 import { InputTypeDefinitionKind } from 'src/engine/api/graphql/workspace-schema-builder/enums/input-type-definition-kind.enum';
 import { ObjectTypeDefinitionKind } from 'src/engine/api/graphql/workspace-schema-builder/enums/object-type-definition-kind.enum';
-import { StoredInputType } from 'src/engine/api/graphql/workspace-schema-builder/factories/composite-input-type-definition.factory';
-import { StoredObjectType } from 'src/engine/api/graphql/workspace-schema-builder/factories/composite-object-type-definition.factory';
-import { StoredEnumGqlType } from 'src/engine/api/graphql/workspace-schema-builder/gql-type-generators/composite-field-enum-type.generator';
 
 // Must be scoped on REQUEST level, because we need to recreate it for each workspaces
 // TODO: Implement properly durable by workspace
