+++ packages/twenty-server/src/engine/api/graphql/workspace-schema-builder/types/graphql-field-config-map.types.ts
@@ -1,0 +1,22 @@
+import {
+  type GraphQLFieldConfigArgumentMap,
+  type GraphQLFieldConfigMap,
+  type GraphQLOutputType,
+} from 'graphql';
+
+export type GraphQLOutputTypeFieldConfigMap = GraphQLFieldConfigMap<
+  string,
+  {
+    type: GraphQLOutputType;
+    description: string | null;
+  }
+>;
+
+export type GraphQLRootTypeFieldConfigMap = GraphQLFieldConfigMap<
+  string,
+  {
+    type: GraphQLOutputType;
+    args: GraphQLFieldConfigArgumentMap;
+    resolve: undefined;
+  }
+>;
